Updated database
YOU WILL PROBABLY NEED TO CLEAR YOUR DATABASE MEASUREMENTS AND COMMAND LOGS
You can do that with these commands in the terminal where your database is:
>>> h5file = open_file(DB_Name, mode="a", title="EARS Measurements Record")
>>> h5file.remove_node("/measurement", recursive=True)
>>> h5file.remove_node("/Logs", recursive=True)
>>> h5file.close()

This changes the command and measurement databases to take time components
instead of a string. It is MUCH easier to query the db and get what you
want back this way.